+++ src/main/java/com/argentinaprograma/tpfinal/repositorios/TecnicoRepositorio.java
@@ -4,7 +4,7 @@ import com.argentinaprograma.tpfinal.dominio.Tecnico;
 
 public interface TecnicoRepositorio {
     
-    Tecnico obtenerTecnicoPorID(int id_tecnico);
+    Tecnico obtenerTecnicoPorID(long id_tecnico);
     void guardarTecnico(Tecnico tecnico);
     void eliminarTecnico(Long id_tecnico);
